Calculate "global" TilingScheme
- move TilingScheme/Negotiator classes to their own module:
  `libertem.io.dataset.base.tiling_scheme`
- Negotiator.get_scheme accepts the DataSet and an approximate Partition
  shape instead of the Partition and a concrete shape
- The TilingScheme is part of UDFParams now
- Move some methods from Partition to DataSet
    - the DataSet now creates the Decoder, as we need it to find out if
      we need to copy
    - move get_base_shape, adjust_tileshape, need_decode
- Add a common method to access DataSet._meta
- More type annotations
